refactoring follow function, moving into separate file and replacing callbacks with promise
